Refactor plotting.
